update version to 3.0.2

+++ cores/esp32/esp32-hal-uart.h
@@ -32,37 +32,43 @@ extern "C" {
 struct uart_struct_t;
 typedef struct uart_struct_t uart_t;
 
-bool _testUartBegin(uint8_t uart_nr, uint32_t baudrate, uint32_t config, int8_t rxPin, int8_t txPin, uint16_t rx_buffer_size, uint16_t tx_buffer_size, bool inverted, uint8_t rxfifo_full_thrhd);
-uart_t* uartBegin(uint8_t uart_nr, uint32_t baudrate, uint32_t config, int8_t rxPin, int8_t txPin, uint16_t rx_buffer_size, uint16_t tx_buffer_size, bool inverted, uint8_t rxfifo_full_thrhd);
+bool _testUartBegin(
+  uint8_t uart_nr, uint32_t baudrate, uint32_t config, int8_t rxPin, int8_t txPin, uint32_t rx_buffer_size, uint32_t tx_buffer_size, bool inverted,
+  uint8_t rxfifo_full_thrhd
+);
+uart_t *uartBegin(
+  uint8_t uart_nr, uint32_t baudrate, uint32_t config, int8_t rxPin, int8_t txPin, uint32_t rx_buffer_size, uint32_t tx_buffer_size, bool inverted,
+  uint8_t rxfifo_full_thrhd
+);
 void uartEnd(uint8_t uart_num);
 
 // This is used to retrieve the Event Queue pointer from a UART IDF Driver in order to allow user to deal with its events
-void uartGetEventQueue(uart_t* uart, QueueHandle_t *q); 
+void uartGetEventQueue(uart_t *uart, QueueHandle_t *q);
 
-uint32_t uartAvailable(uart_t* uart);
-uint32_t uartAvailableForWrite(uart_t* uart);
-size_t uartReadBytes(uart_t* uart, uint8_t *buffer, size_t size, uint32_t timeout_ms);
-uint8_t uartRead(uart_t* uart);
-uint8_t uartPeek(uart_t* uart);
+uint32_t uartAvailable(uart_t *uart);
+uint32_t uartAvailableForWrite(uart_t *uart);
+size_t uartReadBytes(uart_t *uart, uint8_t *buffer, size_t size, uint32_t timeout_ms);
+uint8_t uartRead(uart_t *uart);
+uint8_t uartPeek(uart_t *uart);
 
-void uartWrite(uart_t* uart, uint8_t c);
-void uartWriteBuf(uart_t* uart, const uint8_t * data, size_t len);
+void uartWrite(uart_t *uart, uint8_t c);
+void uartWriteBuf(uart_t *uart, const uint8_t *data, size_t len);
 
-void uartFlush(uart_t* uart);
-void uartFlushTxOnly(uart_t* uart, bool txOnly );
+void uartFlush(uart_t *uart);
+void uartFlushTxOnly(uart_t *uart, bool txOnly);
 
-void uartSetBaudRate(uart_t* uart, uint32_t baud_rate);
-uint32_t uartGetBaudRate(uart_t* uart);
+void uartSetBaudRate(uart_t *uart, uint32_t baud_rate);
+uint32_t uartGetBaudRate(uart_t *uart);
 
-void uartSetRxInvert(uart_t* uart, bool invert);
-bool uartSetRxTimeout(uart_t* uart, uint8_t numSymbTimeout);
-bool uartSetRxFIFOFull(uart_t* uart, uint8_t numBytesFIFOFull);
-void uartSetFastReading(uart_t* uart);
+void uartSetRxInvert(uart_t *uart, bool invert);
+bool uartSetRxTimeout(uart_t *uart, uint8_t numSymbTimeout);
+bool uartSetRxFIFOFull(uart_t *uart, uint8_t numBytesFIFOFull);
+void uartSetFastReading(uart_t *uart);
 
-void uartSetDebug(uart_t* uart);
+void uartSetDebug(uart_t *uart);
 int uartGetDebug();
 
-bool uartIsDriverInstalled(uart_t* uart);
+bool uartIsDriverInstalled(uart_t *uart);
 
 // Negative Pin Number will keep it unmodified, thus this function can set individual pins
 // When pins are changed, it will detach the previous ones
@@ -73,7 +79,6 @@ bool uartSetPins(uint8_t uart_num, int8_t rxPin, int8_t txPin, int8_t ctsPin, in
 int8_t uart_get_RxPin(uint8_t uart_num);
 int8_t uart_get_TxPin(uint8_t uart_num);
 void uart_init_PeriMan(void);
-
 
 // Enables or disables HW Flow Control function -- needs also to set CTS and/or RTS pins
 //    UART_HW_FLOWCTRL_DISABLE = 0x0   disable hardware flow control
@@ -109,7 +114,6 @@ void uart_internal_loopback(uint8_t uartNum, int8_t rxPin);
 void uart_send_break(uint8_t uartNum);
 // Sends a buffer and at the end of the stream, it generates BREAK in the line
 int uart_send_msg_with_break(uint8_t uartNum, uint8_t *msg, size_t msgSize);
-
 
 #ifdef __cplusplus
 }
